+++ cyberdrop_dl/hasher.py
@@ -17,7 +17,7 @@ from cyberdrop_dl.constants import Hashing, TempExt
 if TYPE_CHECKING:
     from yarl import URL
 
-    from cyberdrop_dl.config.config_model import DupeCleanup
+    from cyberdrop_dl.config.settings import DupeCleanup
     from cyberdrop_dl.data_structures.url_objects import MediaItem
     from cyberdrop_dl.managers.manager import Manager
